% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loc.R
\name{plot_loc}
\alias{plot_loc}
\title{Plot locations from database}
\usage{
plot_loc(loc, bbox = c(range(locid$lat), range(locid$lon))[c(1, 3, 2, 4)],
  col = "black")
}
\arguments{
\item{loc}{dataframe with lat and lon (latitude and longitude)}

\item{bbox}{bounding box}

\item{col}{colour for each point (forwarded to points)}
}
\description{
Plot locations from database
}
