% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_per.R
\name{plot_per}
\alias{plot_per}
\title{Plot per time period}
\usage{
plot_per(data, timestamp, type = "hour", ylab = "", main = "")
}
\arguments{
\item{data}{vector with data}

\item{timestamp}{vector with timestamps. Character if type is hour, minute or second, otherwise Date-class}

\item{type}{character, options are: "month", "week", "weekday", "monthday", "day", "hour", "minute", "second"}

\item{ylab}{character, for y axis}

\item{main}{character, title for plot}
}
\value{
plot
}
\description{
Plot per time period
}
