% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loc.R
\name{plot_loc}
\alias{plot_loc}
\title{Plot locations from database}
\usage{
plot_loc(loc, range = c(5.86, 15.05, 47.27, 55.06))
}
\arguments{
\item{loc}{dataframe with lat and lon (latitude and longitude)}

\item{range}{vector of length 4, corners of area to plot}
}
\description{
Plot locations from database
}
